break into command and function
